pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "azstoragetorch"
authors = [
    {name="Microsoft Corporation", email="ascl@microsoft.com"},
]
description = "Azure Storage Connector for PyTorch"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
keywords = ["azure", "pytorch"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "azure-identity<2",
    # This is pinned because the library uses internal APIs of the blob SDK and helps
    # avoid automatically breaking the library if a new version of the SDK breaks the
    # internal API. As new versions of the SDK are released, the ceiling of this version
    # range should be updated to match the new version.
    "azure-storage-blob>=12.24.0,<=12.24.1",
    "torch>=2.6.0,<3",
    "typing-extensions>=4.13.2,<5",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "build",
    "check-manifest",
    "pytest",
    "sphinx",
    "furo",
    "sphinx-copybutton",
    "ruff",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/Azure/azure-storage-for-pytorch"
Issues = "https://github.com/Azure/azure-storage-for-pytorch/issues"
Repository = "https://github.com/Azure/azure-storage-for-pytorch"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "azstoragetorch._version.__version__"}
